Add initial `README.md`, update platform target configurations, and fix `pwd()` implementation for native platforms.

+++ ktkit/src/nativeMain/kotlin/io/github/smyrgeorge/ktkit/util/process.native.kt
@@ -28,4 +28,4 @@ actual fun registerShutdownHook() {
 }
 
 @OptIn(ExperimentalForeignApi::class)
-actual fun pwd(): String = ByteArray(1024).usePinned { getcwd(it.addressOf(0), 1024u) }!!.toKString()
+actual fun pwd(): String = ByteArray(1024).usePinned { getcwd(it.addressOf(0), 1024.toULong()) }!!.toKString()
